ESOLA-238 Host Collection ib_iser vars folder missing
Because the "vars" folder under the ib_iser role is missing, the automation
failed to find the path speficied in iface.yml > "ansible.builtin.include_vars".
The vars folder is where Ansible supposed to start the lookup but since it's
missing, Ansible uses other paths as starting point which caused the failure.
The fix is to add the vars as empty folder.