Provide documentation for known issues related to iSER and selinux.
When running the automation with enforcing first then permissive selinux policy,
it is possible that the IB links will still not become Active. The fix is to provide
a manual workaround/solution for customers. Usually they should have set to permissive
initially to get iSER working.
Remove mentions of using permissive and debug issue as we don't have a guidance on how to debug.