Add checks for mstconfig in ib role firmware tasks.
Currently, reglardless if the mstconfig options are set in the xxx_ib_interfaces
variable, the firmware tasks for InfiniBand are executed. This was not
the intended design and was harmless until we have to support virtual
functions where the InfiniBand ports are virtualized and those mstconfig
commands do not work.

The changes are to check for mstconfig options and only run those
firmware tasks when the options are present. This is the intended design
and resolve the virtualized InfiniBand ports issue.